+++ ICD.Connect.Krang.CrestronSPlus/Properties/AssemblyInfo.cs
@@ -4,5 +4,4 @@
 [assembly: AssemblyCompany("ICD Systems")]
 [assembly: AssemblyProduct("ICD.Connect.Krang.CrestronSPlus")]
 [assembly: AssemblyCopyright("Copyright © ICD Systems 2021")]
-[assembly: AssemblyVersion("15.3.0.0")]
-
+[assembly: AssemblyVersion("16.0.0.0")]