+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Changed
  - Added a firmware column to the controlsystem health console table
+ - Generated destination groups are now given uuids
 
 ## [15.3.0] - 2021-01-14
 ### Added